+++ config/urls.py
@@ -9,4 +9,5 @@ urlpatterns = [
     path('api/company/', include('apps.api.urls.company_urls', namespace='company_endpoints')),
     path('api/branch/', include('apps.api.urls.branch_urls', namespace='branch_endpoints')),
     path('api/telebot/', include('apps.api.urls.telebot_urls', namespace='telebot_endpoints')),
+    path('api/connect/', include('apps.api.urls.connect_urls', namespace='connect_endpoints')),
 ]